[l2k-server-#95] Usable items - l2day buff scrolls and scrolls of escape - dependency updates

+++ l2k-server-game/src/main/kotlin/org/l2kserver/game/handler/dto/response/CharacterListResponse.kt
@@ -104,8 +104,8 @@ data class CharacterListResponse(
             putInt(it.hairColor)
             putInt(it.faceType)
 
-            putDouble(it.stats.maxHp.toDouble())
-            putDouble(it.stats.maxMp.toDouble())
+            putDouble(it.stats.maxHp)
+            putDouble(it.stats.maxMp)
 
             putInt(
                 if (it.deletionDate == null) 0
